pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bugraid-itops-agent"
version = "0.1.0"
description = "BugRaid ITOps Research Agent — deep research for production incidents"
requires-python = ">=3.10"
dependencies = [
    # Orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",

    # LLM
    "anthropic>=0.30.0",

    # Embeddings (local, no API cost)
    "sentence-transformers>=3.0.0",

    # Vector store + graph
    "opensearch-py>=2.6.0",
    "neo4j>=5.20.0",

    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",

    # Config + validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Utils
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "tenacity>=8.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",  # for FastAPI test client
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
